Fix NewMaxFee() truncating int64 fee-bump fee to uint32
FeeBumpTransaction.Fee is XDR Int64 but NewMaxFee() was casting it to
uint32, silently discarding the upper 32 bits. Any fee-bump fee above
~429.5 XLM (4,294,967,295 stroops) returned a wrong value.

Change return type from uint32 to int64 — this is a breaking change.

Fixes #5931

+++ ingest/ledger_transaction.go
@@ -736,12 +736,12 @@ func (t *LedgerTransaction) InnerTransactionHash() (string, bool) {
 	return hex.EncodeToString(innerHash[:]), true
 }
 
-func (t *LedgerTransaction) NewMaxFee() (uint32, bool) {
+func (t *LedgerTransaction) NewMaxFee() (int64, bool) {
 	if !t.Envelope.IsFeeBump() {
 		return 0, false
 	}
 
-	return uint32(t.Envelope.FeeBumpFee()), true
+	return t.Envelope.FeeBumpFee(), true
 }
 
 func (t *LedgerTransaction) Successful() bool {
